+++ demisto_sdk/commands/common/errors.py
@@ -973,6 +973,11 @@ ERROR_CODE = {
         "ui_applicable": False,
         "related_field": "",
     },
+    "layout_container_contains_invalid_types": {
+        "code": "LO107",
+        "ui_applicable": False,
+        "related_field": "",
+    },
     # MP - Mappers
     "invalid_from_version_in_mapper": {
         "code": "MP100",
@@ -3245,6 +3250,14 @@ class Errors:
     @error_code_decorator
     def layout_non_existent_script_id(layout, scripts):
         return f"In layout {layout} the following scripts were not found in the id_set.json file: {scripts}"
+
+    @staticmethod
+    @error_code_decorator
+    def layout_container_contains_invalid_types(invalid_types):
+        return (
+            f"The following invalid types were found in the layout: {str(invalid_types)}. Those types are not"
+            f" supported in XSIAM, remove them or change the layout to be XSOAR only"
+        )
 
     @staticmethod
     def suggest_fix_non_existent_script_id() -> str: